Fix `current_time` handling for legacy timestamp semantics
Correct current_time behavior in legacy timestamp semantics when the session time zone’s current offset differs from its offset on 1970-01-01.
